fix(i18n): continue on warning
Right now the warnings don't just point a potential issue out and continue as usual, but prevent the i18n base URL to be set. I think that the warnings should either be an error so that one is more likely to guess an influence on behavior or the warnings should not influence the behavior. This implements the latter option.